batch size db

+++ dlrm_s_pytorch_pseudo_cpustb.py
@@ -987,7 +987,7 @@ def main():
     args = parser.parse_args() 
     args.world_size = 1 
 
-    args.mini_batch_size = args.mini_batch_size / args.number_of_cpu_node 
+    args.mini_batch_size = int(args.mini_batch_size / args.number_of_cpu_node) 
     
     if args.test_mini_batch_size < 0:
         # if the parameter is not set, use the training batch size
